Flesh out api for forbidding patterns from cells
This also corrects a bug where the wave wasn't reset when a
contradiction was reached while using the "retry" api. Custom logic for
preventing patterns from appearing in cells must now be encapsulated in
an implementation of the `ForbidPattern` trait so it can be invoked each
time the wave is reset upon contradiction.